Updated IInputSystem

+++ SimpleSongs/SimpleSongsView/Interfaces/IInputSystem.cs
@@ -9,6 +9,6 @@ namespace SimpleSongsView.Interfaces
     public interface IInputSystem
     {
         public string FetchStringValue(string prompt);
-        public decimal FetchDecimalValue(string prompt);
+        public double FetchDecimalValue(string prompt);
     }
 }
